port tests to connection_string.ts and add MongoCredentials validation in testing
