+++ test/unit/connection_string.test.ts
@@ -5,12 +5,328 @@ import { promisify } from 'util';
 
 import { MongoCredentials } from '../../src/cmap/auth/mongo_credentials';
 import { $EXTERNAL_AUTH_SOURCE_MECHANISMS, AuthMechanism } from '../../src/cmap/auth/providers';
-import { resolveSRVRecord } from '../../src/connection_string';
+import { parseOptions, resolveSRVRecord } from '../../src/connection_string';
+import { MongoDriverError, MongoInvalidArgumentError, MongoParseError } from '../../src/error';
+import { MongoOptions } from '../../src/mongo_client';
+import { loadSpecTests } from '../spec';
 
-describe('Connection String', () => {
+// NOTE: These are cases we could never check for unless we write our own
+//       url parser. The node parser simply won't let these through, so we
+//       are safe skipping them.
+const skipTests = [
+  'Invalid port (negative number) with hostname',
+  'Invalid port (non-numeric string) with hostname',
+  'Missing delimiting slash between hosts and options',
+
+  // These tests are only relevant to the native driver which
+  // cares about specific keys, and validating their values
+  'Unrecognized option keys are ignored',
+  'Unsupported option values are ignored',
+
+  // We don't actually support `wtimeoutMS` which this test depends upon
+  'Deprecated (or unknown) options are ignored if replacement exists'
+];
+
+describe('Connection String', function () {
+  it('should not support auth passed with user', function () {
+    const optionsWithUser = {
+      authMechanism: 'SCRAM-SHA-1',
+      auth: { user: 'testing', password: 'llamas' }
+    };
+
+    expect(() => parseOptions('mongodb://localhost', optionsWithUser as any)).to.throw(
+      MongoParseError
+    );
+  });
+
+  it('should support auth passed with username', function () {
+    const optionsWithUsername = {
+      authMechanism: 'SCRAM-SHA-1',
+      auth: { username: 'testing', password: 'llamas' }
+    };
+    const options = parseOptions('mongodb://localhost', optionsWithUsername as any);
+    expect(options.credentials).to.containSubset({
+      source: 'admin',
+      username: 'testing',
+      password: 'llamas'
+    });
+  });
+
+  it('should provide a default port if one is not provided', function () {
+    const options = parseOptions('mongodb://hostname');
+    expect(options.hosts[0].socketPath).to.be.undefined;
+    expect(options.hosts[0].host).to.be.a('string');
+    expect(options.hosts[0].port).to.equal(27017);
+  });
+
+  it('should parse multiple readPreferenceTags', function () {
+    const options = parseOptions(
+      'mongodb://hostname?readPreferenceTags=bar:foo&readPreferenceTags=baz:bar'
+    );
+    expect(options.readPreference.tags).to.deep.equal([{ bar: 'foo' }, { baz: 'bar' }]);
+  });
+
+  it('should parse boolean values', function () {
+    let options = parseOptions('mongodb://hostname?retryWrites=1');
+    expect(options.retryWrites).to.equal(true);
+    options = parseOptions('mongodb://hostname?retryWrites=false');
+    expect(options.retryWrites).to.equal(false);
+    options = parseOptions('mongodb://hostname?retryWrites=t');
+    expect(options.retryWrites).to.equal(true);
+  });
+
+  it('should parse compression options', function () {
+    const options = parseOptions('mongodb://localhost/?compressors=zlib&zlibCompressionLevel=4');
+    expect(options).to.have.property('compressors');
+    expect(options.compressors).to.include('zlib');
+    expect(options.zlibCompressionLevel).to.equal(4);
+  });
+
+  it('should parse `readConcernLevel`', function () {
+    const options = parseOptions('mongodb://localhost/?readConcernLevel=local');
+    expect(options).to.have.property('readConcern');
+    expect(options.readConcern.level).to.equal('local');
+  });
+
+  it('should parse `authMechanismProperties`', function () {
+    const options = parseOptions(
+      'mongodb://user%40EXAMPLE.COM:secret@localhost/?authMechanismProperties=SERVICE_NAME:other,SERVICE_REALM:blah,CANONICALIZE_HOST_NAME:true&authMechanism=GSSAPI'
+    );
+    expect(options.credentials.mechanismProperties).to.deep.include({
+      SERVICE_NAME: 'other',
+      SERVICE_REALM: 'blah',
+      CANONICALIZE_HOST_NAME: true
+    });
+    expect(options.credentials.mechanism).to.equal(AuthMechanism.MONGODB_GSSAPI);
+  });
+
+  it('should provide default authSource when valid AuthMechanism provided', function () {
+    const options = parseOptions(
+      'mongodb+srv://jira-sync.pw0q4.mongodb.net/testDB?authMechanism=MONGODB-AWS&retryWrites=true&w=majority'
+    );
+    expect(options.credentials.source).to.equal('$external');
+  });
+
+  it('should parse a numeric authSource with variable width', function () {
+    const options = parseOptions('mongodb://test@localhost/?authSource=0001');
+    expect(options.credentials.source).to.equal('0001');
+  });
+
+  it('should not remove dbName from the options if authSource is provided', function () {
+    const dbName = 'my-db-name';
+    const authSource = 'admin';
+    const options = parseOptions(
+      `mongodb://myName:myPassword@localhost:27017/${dbName}?authSource=${authSource}`
+    );
+
+    expect(options).has.property('dbName', dbName);
+    expect(options.credentials).to.have.property('source', authSource);
+  });
+
+  it('should parse a replicaSet with a leading number', function () {
+    const options = parseOptions('mongodb://localhost/?replicaSet=123abc');
+    expect(options).to.have.property('replicaSet');
+    expect(options.replicaSet).to.equal('123abc');
+  });
+
+  context('when both tls and ssl options are provided', function () {
+    context('when the options are provided in the URI', function () {
+      context('when the options are equal', function () {
+        context('when both options are true', function () {
+          const options = parseOptions('mongodb://localhost/?tls=true&ssl=true');
+
+          it('sets the tls option', function () {
+            expect(options.tls).to.be.true;
+          });
+
+          it('does not set the ssl option', function () {
+            expect(options).to.not.have.property('ssl');
+          });
+        });
+
+        context('when both options are false', function () {
+          const options = parseOptions('mongodb://localhost/?tls=false&ssl=false');
+
+          it('sets the tls option', function () {
+            expect(options.tls).to.be.false;
+          });
+
+          it('does not set the ssl option', function () {
+            expect(options).to.not.have.property('ssl');
+          });
+        });
+      });
+
+      context('when the options are not equal', function () {
+        it('raises an error', function () {
+          expect(() => {
+            parseOptions('mongodb://localhost/?tls=true&ssl=false');
+          }).to.throw(MongoParseError, 'All values of tls/ssl must be the same.');
+        });
+      });
+    });
+
+    context('when the options are provided in the options', function () {
+      context('when the options are equal', function () {
+        context('when both options are true', function () {
+          const options = parseOptions('mongodb://localhost/', { tls: true, ssl: true });
+
+          it('sets the tls option', function () {
+            expect(options.tls).to.be.true;
+          });
+
+          it('does not set the ssl option', function () {
+            expect(options).to.not.have.property('ssl');
+          });
+        });
+
+        context('when both options are false', function () {
+          context('when the URI is an SRV URI', function () {
+            const options = parseOptions('mongodb+srv://localhost/', { tls: false, ssl: false });
+
+            it('overrides the tls option', function () {
+              expect(options.tls).to.be.false;
+            });
+
+            it('does not set the ssl option', function () {
+              expect(options).to.not.have.property('ssl');
+            });
+          });
+
+          context('when the URI is not SRV', function () {
+            const options = parseOptions('mongodb://localhost/', { tls: false, ssl: false });
+
+            it('sets the tls option', function () {
+              expect(options.tls).to.be.false;
+            });
+
+            it('does not set the ssl option', function () {
+              expect(options).to.not.have.property('ssl');
+            });
+          });
+        });
+      });
+
+      context('when the options are not equal', function () {
+        it('raises an error', function () {
+          expect(() => {
+            parseOptions('mongodb://localhost/', { tls: true, ssl: false });
+          }).to.throw(MongoParseError, 'All values of tls/ssl must be the same.');
+        });
+      });
+    });
+  });
+
+  describe('validation', function () {
+    it('should validate compressors options', function () {
+      expect(() => parseOptions('mongodb://localhost/?compressors=bunnies')).to.throw(
+        MongoInvalidArgumentError,
+        'bunnies is not a valid compression mechanism'
+      );
+    });
+
+    it('should validate authMechanism', function () {
+      expect(() => parseOptions('mongodb://localhost/?authMechanism=DOGS')).to.throw(
+        MongoParseError,
+        'authMechanism one of MONGODB-AWS,MONGODB-CR,DEFAULT,GSSAPI,PLAIN,SCRAM-SHA-1,SCRAM-SHA-256,MONGODB-X509, got DOGS'
+      );
+    });
+
+    it('should validate readPreference', function () {
+      expect(() => parseOptions('mongodb://localhost/?readPreference=llamasPreferred')).to.throw(
+        MongoDriverError, // not parse Error b/c thrown from ReadPreference construction
+        'Invalid read preference mode "llamasPreferred"'
+      );
+    });
+  });
+
+  describe('spec tests', function () {
+    const suites = loadSpecTests('connection-string').concat(loadSpecTests('auth'));
+
+    for (const suite of suites) {
+      describe(suite.name, function () {
+        for (const test of suite.tests) {
+          it(`${test.description}`, function () {
+            if (skipTests.includes(test.description)) {
+              return this.skip();
+            }
+
+            const message = `"${test.uri}"`;
+
+            const valid = test.valid;
+            if (valid) {
+              const options = parseOptions(test.uri);
+              expect(options, message).to.be.ok;
+
+              if (test.hosts) {
+                for (const [index, { host, port }] of test.hosts.entries()) {
+                  expect(options.hosts[index], message).to.satisfy(e => {
+                    return e.host === host || e.socketPath === host;
+                  });
+                  if (typeof port === 'number') expect(options.hosts[index].port).to.equal(port);
+                }
+              }
+
+              if (test.auth && test.auth.db != null) {
+                expect(options.dbName, message).to.equal(test.auth.db);
+              }
+
+              if (test.auth && test.auth.username) {
+                expect(options.credentials, message).to.exist;
+
+                if (test.auth.db != null) {
+                  expect(options.credentials.source, message).to.equal(test.auth.db);
+                }
+
+                if (test.auth.username != null) {
+                  expect(options.credentials.username, message).to.equal(test.auth.username);
+                }
+
+                if (test.auth.password != null) {
+                  expect(options.credentials.password, message).to.equal(test.auth.password);
+                }
+              }
+
+              if (test.options) {
+                for (const [optionKey, optionValue] of Object.entries(test.options)) {
+                  switch (optionKey) {
+                    case 'authmechanism':
+                      expect(options.credentials.mechanism, message).to.eq(optionValue);
+                      break;
+                    case 'authmechanismproperties':
+                      expect(options.credentials.mechanismProperties, message).to.deep.eq(
+                        optionValue
+                      );
+                      break;
+                    case 'replicaset':
+                      expect(options.replicaSet, message).to.equal(optionValue);
+                      break;
+                    case 'w':
+                      expect(options.writeConcern.w).to.equal(optionValue);
+                      break;
+                    default:
+                      throw Error(`This options is not covered by the spec test: ${optionKey}`);
+                  }
+                }
+              }
+            } else {
+              expect(() => parseOptions(test.uri), message).to.throw();
+            }
+          });
+        }
+      });
+    }
+  });
+
+  describe('mongodb+srv', function () {
+    it('should parse a default database', function () {
+      const options = parseOptions('mongodb+srv://test1.test.build.10gen.cc/somedb');
+      expect(options.dbName).to.equal('somedb');
+      expect(options.srvHost).to.equal('test1.test.build.10gen.cc');
+    });
+  });
+
   describe('resolveSRVRecord()', () => {
-    let resolveSrvStub: sinon.SinonStub;
-    let resolveTxtStub: sinon.SinonStub;
     const resolveSRVRecordAsync = promisify(resolveSRVRecord);
 
     afterEach(async () => {
@@ -29,34 +345,38 @@ describe('Connection String', () => {
 
       const mockRecord: string[][] = [[txtRecord]];
 
-      // first call is for the driver initial connection
-      // second call will check the poller
-      resolveSrvStub = sinon.stub(dns, 'resolveSrv').callsFake((address, callback) => {
+      // first call is for stubbing resolveSrv
+      // second call is for stubbing resolveTxt
+      sinon.stub(dns, 'resolveSrv').callsFake((address, callback) => {
         return process.nextTick(callback, null, mockAddress);
       });
 
-      resolveTxtStub = sinon.stub(dns, 'resolveTxt').callsFake((address, thisIsWhatWeAreTesting) => {
-        thisIsWhatWeAreTesting(null, mockRecord);
+      sinon.stub(dns, 'resolveTxt').callsFake((address, whatWeTest) => {
+        whatWeTest(null, mockRecord);
       });
     }
 
     for (const mechanism of $EXTERNAL_AUTH_SOURCE_MECHANISMS) {
       it(`should set authSource to $external for ${mechanism} external mechanism`, async function () {
         makeStub('authSource=thisShouldNotBeAuthSource');
+        const credentials = new MongoCredentials({
+          source: '$external',
+          mechanism,
+          username: 'username',
+          password: mechanism === AuthMechanism.MONGODB_X509 ? undefined : 'password',
+          mechanismProperties: {}
+        });
+        credentials.validate();
+
         const options = {
-          credentials: new MongoCredentials({
-            source: '$external',
-            mechanism: mechanism,
-            username: 'username',
-            password: 'password',
-            mechanismProperties: {}
-          }),
+          credentials,
           srvHost: 'test.mock.test.build.10gen.cc',
           srvServiceName: 'mongodb',
           userSpecifiedAuthSource: false
-        };
+        } as MongoOptions;
 
-        await resolveSRVRecordAsync(options as any);
+        await resolveSRVRecordAsync(options);
+        expect(options.credentials === credentials).to.be.true;
         expect(options).to.have.nested.property('credentials.source', '$external');
       });
     }
@@ -64,39 +384,47 @@ describe('Connection String', () => {
     it('should set a default authSource for non-external mechanisms with no user-specified source', async function () {
       makeStub('authSource=thisShouldBeAuthSource');
 
+      const credentials = new MongoCredentials({
+        source: 'admin',
+        mechanism: AuthMechanism.MONGODB_SCRAM_SHA256,
+        username: 'username',
+        password: 'password',
+        mechanismProperties: {}
+      });
+      credentials.validate();
+
       const options = {
-        credentials: new MongoCredentials({
-          source: 'admin',
-          mechanism: AuthMechanism.MONGODB_SCRAM_SHA256,
-          username: 'username',
-          password: 'password',
-          mechanismProperties: {}
-        }),
+        credentials,
         srvHost: 'test.mock.test.build.10gen.cc',
         srvServiceName: 'mongodb',
         userSpecifiedAuthSource: false
-      };
+      } as MongoOptions;
 
-      await resolveSRVRecordAsync(options as any);
+      await resolveSRVRecordAsync(options);
+      expect(options.credentials !== credentials).to.be.true;
       expect(options).to.have.nested.property('credentials.source', 'thisShouldBeAuthSource');
     });
 
     it('should retain credentials for any mechanism with no user-sepcificed source and no source in DNS', async function () {
       makeStub('');
+      const credentials = new MongoCredentials({
+        source: 'admin',
+        mechanism: AuthMechanism.MONGODB_SCRAM_SHA256,
+        username: 'username',
+        password: 'password',
+        mechanismProperties: {}
+      });
+      credentials.validate();
+
       const options = {
-        credentials: new MongoCredentials({
-          source: 'admin',
-          mechanism: AuthMechanism.MONGODB_SCRAM_SHA256,
-          username: 'username',
-          password: 'password',
-          mechanismProperties: {}
-        }),
+        credentials,
         srvHost: 'test.mock.test.build.10gen.cc',
         srvServiceName: 'mongodb',
         userSpecifiedAuthSource: false
-      };
+      } as MongoOptions;
 
       await resolveSRVRecordAsync(options as any);
+      expect(options.credentials === credentials).to.be.true;
       expect(options).to.have.nested.property('credentials.source', 'admin');
     });
   });
